# Optimize mocking further by retaining mocks across sessions