This takes the optimizations a bit further by retaining mocked modules between tests in a `?WITH_MOCKED_SETUP`. That is, between such tests, the mocked modules are not restored to original, but instead the modules are left in a mocked state, the next test will need them mocked.

For a fairly large unit test module with 44 test cases with many mocked functions (the same as in #18) the timing now looks like in the table below (lower value is better).

Execution time  | Code
----------------|----------------
67s             | Original
32s             | PR #18
9s              | This PR

This PR currently shares its first commits with that of #18, I didn't know how to do it before that one gets merged, yet I thought it may be worth to make a PR so any discussions can happen. I can rebase this later on when it is suitable. I suppose it may make sense to have this further optimization as a separate PR instead of baking it into the #18, but I'm open to any approach, just let me know.

The general optimization idea in this PR is as follows:

* Extend [`mockgyver:exec/3→4`](https://github.com/tomas-abrahamsson/mockgyver/blob/retain-mocks-across-sessions/src/mockgyver.erl#L438-L462) with an extra parameter for options
* In `?WITH_MOCKED_SETUP`, set a `mock_sequence` option to tell mockgyver that the session is part of a longer mock sequence.
* When a session in a sequence ends, don't restore mocked modules, instead await the next session.
* But when the _last_ session in the sequence ends, _do_ restore mocked modules.
* When a session in a sequence starts, don't re-setup mocked modules.
* Use a test signature to tell different mock sequences apart. This signature consists of the current location and some more to make it unique for each `?WITH_MOCKED_SETUP`.
* When awaiting the next sequence, if it doesn't started quickly enough, currently 100ms, then restore the mocked modules. Something might be wrong. If we this turned out to be overly cautious, the only thing that happened was that the optimization didn't kick in. It might kick in for the next session.
* With just `?MOCK` as opposed to `?WITH_MOCKED_SETUP`, don't set any mock sequence options.
* There is a `no_mock_sequence` option to disable this optimization.

So modules stay mocked even after a `mockgyver:exec` session. But only between test in `?WITH_MOCKED_SETUP`. And with a safety timer.

This PR introduces a backwards incompatible change: it no longer works to mock the `application` module with `?WITH_MOCKED_SETUP`, since the `mockgyver:exec/4` depends on it (uses it). The retaining of mockings between sessions collides with possibilities of mocking the `application` module. I figure it is a not so good practice to mock modules central to Erlang/OTP, so I hope this is not too big of a draw-back. (I don't think it will work even if the mocking is done carefully to fall back to the renamed original: the mocking module requires a session but when calling `mockgyver:exec/4`, we are between sessions. The `no_mock_sequence` option may be used though.)

Of course, code review and comments are welcome.

**Edit:** Forgot to say that I have tested this in a large project. Only small adjustments were needed. For instance one test didn't properly wait for a process to terminate, and since the tests now run faster, there was a `{error,{already_started,<pid>}}` false positive.