@tomas-abrahamsson 2021-08-12
I added a commit, to provide a better error message to the user/programmer in case some central module is mocked, in particular something that is called by eunit or mockgyver itself between sessions in a sequence. The error message recomments trying to avoid mocking that module, or else to use the `no_mock_sequence` option to the `?WITH_MOCKED_SETUP`.