@tomas-abrahamsson 2021-08-12
I force-pushed an extra line of unit test. Now it is ready for review. Sorry for the noise. (I intended to start another review for this, just to make it clear what I was about to do, but messed up, and in the end, it just got force-pushed.)