@tomas-abrahamsson 2021-08-16
Hmm.. revisiting again. In another corpus of unit tests, I've spotted mockings of `filelib` and `timer`, both of which could potentially be called by eunit between unit tests in a generator, meaning letting mocking modules stay loaded could prove problematic. 

Maybe need to implement some auto-disabling of the optimization depending on what modules are mocked to make sure existing tests still work. (Still warranted, I should think, given the speedup(?))