@tomas-abrahamsson 2021-08-24
I just force-pushed a change to the last commit, and I think it is (really) ready for review now. (I will also try to run some more extensive tests in a big project, meanwhile.)

The situation with mocked functions possibly getting called between unit tests (by eunit, mockgyver or the rest of the runtime) was bugging me, but I think I found a nice and clean solution for it now: In the `await_next_session` state, just return `{ok,undefined}` to any `reg_call_and_get_action`, this will in effect redirect the call to the renamed `module^` and the original code will get invoked.

(I've kept a dead-end on the `https://github.com/tomas-abrahamsson/mockgyver/tree/cul-de-sac/disable-mock-seq-optimization` in the dubious case it will still be needed. On this branch, it basically disables the mock-sequence optimization if any module of any of the `kernel`, `stdlib` or `sasl` apps is mocked, and also prints a warning about it. I will delete this branch in a while if will not be needed.)