@klajo 2021-10-02
Marking this as closed, in favor of #21. Thanks for your hard work.